requirements.txt
Flask==2.3.3
Flask-PyMongo==2.3.0
Flask-Login==0.6.2
python-dotenv==1.0.0
pymongo[srv]==4.4.1
bcrypt==4.0.1